Keep cursor position while saving
Saving a file triggers formatting the file in a certain way. The most
basic form of which is ensuring there's a final newline. When that
happens, the cursor jumps to that inserted newline, which can be
annoying when you're writing in a file and alt-tabbing from zed to
something else (assuming you have autosave: on_window_change).

This change saves the position of the cursor so this operation works but
attempts to keep your cursor where it was, so you can keep typing. It's
the same story with invoking save yourself when your cursor it at the
end of a file.